+++ Contoso.Test.Flow.Test/Constants.cs
@@ -4,7 +4,7 @@
     {
         public const string FilterResultsFolder = "FilterResults";
         public const string TargetFilterFolder = "TargetFilter";
-        public const string ProjectFileFullPath = @"C:\LogicBuilder\FlowProjects\Contoso.Test\Contoso.Test.lbproj";
-        public const string ShopBslProjectFileFullPath = @"C:\LogicBuilder\FlowProjects\Shop.Bsl\Shop.Bsl.lbproj";
+        public const string ProjectFileFullPath = @"C:\TelerikLogicBuilder\FlowProjects\Contoso.Test\Contoso.Test.lbproj";
+        public const string ShopBslProjectFileFullPath = @"C:\TelerikLogicBuilder\FlowProjects\Shop.Bsl\Shop.Bsl.lbproj";
     }
 }