+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* 2026-09-05 - AB#219: Test Coverage for XML validation on single blank exit.
+* 2026-09-05 - AB#219: Allow XML validation to run for single blank exit connector.
 * 2026-09-03 - AB#221: Fix bug in ConstructorXmlBuilder logic.
 * 2026-09-02 - AB#221: Removed unused class members form ConstructorXmlBuilder.
 * 2026-09-01 - AB#221: Removed model specific XML generation.